// resources/js/types/pagination.d.ts
export interface Pagination<T> {
    data: T[];
    current_page: number;
    last_page: number;
    per_page: number;
    total: number;
    from: number;
    to: number;
    links: {
        url: string | null;
        label: string;
        active: boolean;
    }[];
    meta: any;
    perPage: number; // Added to match TeacherCreatedProps
}